WIP: T03 -- Delete dead code and fix misplaced fixtures
