+++ src/hassette/test_utils/helpers.py
@@ -405,11 +405,6 @@ class {class_name}Config(AppConfig):
     app_file.write_text(textwrap.dedent(content).lstrip())
 
 
-async def emit_service_event(hassette: "Hassette", event: "HassetteServiceEvent") -> None:
-    """Inject a HassetteServiceEvent into the bus."""
-    await hassette.send_event(event)
-
-
 async def emit_file_change_event(hassette: "Hassette", changed_paths: set[Path]) -> None:
     """Emit a synthetic file-watcher event for the given paths."""
     event = HassetteFileWatcherEvent.create_event(changed_file_paths=changed_paths)
